Add a port number to broadcast.
Before the broadcast did not have a specific port and the implementation
chose a random one. This is obviously pretty dumb since the broadcast
must also be received by someone. If it doesn't have a fixed port, the
client would be required to watch ALL packets on the network, which is,
not ideal to say the least.

Signed-off-by: xSlendiX <slendi@socopon.com>