Add and implement `INFO` command.
This commit implements the INFO command which is used to give the
client infromation about the robot, such as drive mode or what features
are enabled (currently only VideoStreaming available). More info about
this command in the specification file. Besides this, this commit also
fixed a lot of critical bugs that went unnoticed in the previous
commits.

Signed-off-by: xSlendiX <slendi@socopon.com>