Implement password hashing
This commit is great for security, since instead of storing the password
as a string inside the binary, it is now read from a file, hashed using
openssl.

This is not complete though -- since it requires setup. This commit
assumes a 'psk' file is found and contains the sha256 of the password,
as bytes (not hexadecimal representation).

Signed-off-by: xSlendiX <slendi@socopon.com>