Add a way to change log level and fix socket not being closed
This patch adds a way to change the log level either by using the -l
command line option or by using the environment variable "LOG_LEVEL".

Additionally, this patch also makes it so that, when broadcasting, the
socket gets closed. A future improvement can be made though, albeit not
important: keep the socket open all time while broadcasting is enabled,
this way you wouldn't need to create a new socket everytime and close
it.

Signed-off-by: xSlendiX <slendi@socopon.com>